Refactored code

+++ src/main/java/inventorymanagementserver/item/ItemServiceImpl.java
@@ -1,5 +1,8 @@
 package inventorymanagementserver.item;
 
+import inventorymanagementserver.category.Category;
+import inventorymanagementserver.category.CategoryRepository;
+import inventorymanagementserver.exception.InventoryException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -9,59 +12,68 @@ import java.util.List;
 public class ItemServiceImpl implements ItemService {
 
     @Autowired
-    ItemRepository itemRepository;
+    private ItemRepository itemRepository;
     @Autowired
-    ItemCategoryFacade itemCategoryFacade;
+    private CategoryRepository categoryRepository;
 
     @Override
-    public Item registerItem(Item item) throws Exception {
-        Item existingItem = itemRepository.findByName(item.getName());
-        if(existingItem != null){
-            throw new Exception("Item with the a same name already exist");
+    public Item registerItem(Item item) {
+        Item existingItemWithSameName = itemRepository.findByName(item.getName());
+        if(existingItemWithSameName != null){
+            throw new InventoryException("Item with the a same name already exist");
         }
         return itemRepository.save(item);
     }
 
     @Override
-    public Item findById(Long id) throws Exception {
+    public Item findById(Long id) {
         String message = String.format("Item with the ID %s not found", id);
         return itemRepository.findById(id)
-                .orElseThrow(()-> new Exception(message));
+                .orElseThrow(()-> new InventoryException(message));
     }
 
     @Override
-    public Item findByName(String name) throws Exception {
+    public Item findByName(String name) {
         Item item = itemRepository.findByName(name);
         if(item == null){
             String message = String.format("Item with name %s not found", name);
-            throw new Exception(message);
+            throw new InventoryException(message);
         }
         return item;
     }
 
     @Override
-    public List<Item> findAll() throws Exception {
+    public List<Item> findAll() {
         List<Item> items = itemRepository.findAll();
-        if(items == null || items.isEmpty()){
-            throw new Exception("No item/s found");
+        if(items.isEmpty()){
+            throw new InventoryException("No item/s found");
         }
         return items;
     }
 
     @Override
-    public Item updateItem(Item item) throws Exception {
+    public Item updateItem(Item item) {
+        checkIfCategoryExist(item.getCategory());
         Item existingItemWithSameName = itemRepository.findByName(item.getName());
         if(existingItemWithSameName != null && !existingItemWithSameName.equals(item)){
             String message = String.format("Item name is already taken");
-            throw new Exception(message);
+            throw new InventoryException(message);
         }
-        itemCategoryFacade.checkIfCategoryExist(item.getCategory());
         return itemRepository.save(item);
     }
 
     @Override
-    public void deleteById(Long id) throws Exception {
-        findById(id);
+    public void deleteById(Long id) {
+        if(!itemRepository.existsById(id)){
+            String message = String.format("Item with ID %s not found", id);
+            throw new InventoryException(message);
+        }
         itemRepository.deleteById(id);
+    }
+
+    private void checkIfCategoryExist(Category category) {
+        if(!categoryRepository.existsById(category.getId())){
+            throw new InventoryException("Category no longer exist");
+        }
     }
 }
